Used methods to convert between sql date format and Calendar Obj where required

+++ app/src/main/java/com/example/james/todolist/model/Task.java
@@ -24,7 +24,7 @@ public class Task implements Serializable{
 
     private static String invalidDateErrorMessage = "Invalid date entered: day - %s, month - %s, year - %s";
 
-    public Task(int id, String outline, String extraDetails, Calendar creationDate, Calendar dueDate, boolean status){
+    public Task(long id, String outline, String extraDetails, Calendar creationDate, Calendar dueDate, boolean status){
         this.id = id;
         this.outline = outline;
         this.extraDetails = extraDetails;
